refactor: move `src/yarn.ts` to `src/yarn/index.ts`
